imooc/requirements.txt
celery==
croniter==
django-celery==3.1.17
django-celery-beat==
kombu==